feat(npm)!: rename @vizij/arora-web-wasm to @vizij/runtime
The browser wrapper's package name is its client surface, so it names what
the client gets — the Vizij runtime, pairing with @vizij/runtime-react —
not the implementation choice (arora) or the compilation detail (wasm).
Only the package identity changes: the vizij-arora-web crate, the wasm-pack
artifacts, and the JS class names all keep their names, and the wasm-build
scripts stay named for the crate they build.

@vizij/runtime starts at 1.0.0; earlier versions were published as
@vizij/arora-web-wasm.

Part of the arora 9 wave migration
(https://linear.app/semio-ai/issue/VIZ-53).
