+++ npm/@vizij/animation-module/README.md
@@ -1,9 +1,9 @@
 # @vizij/animation-module
 
-The vizij animation engine ([`vizij-animation-core`](https://github.com/vizij-ai/vizij-rs/tree/main/crates/animation/vizij-animation-core)) packaged as an **Arora wasm module**, shipped as importable assets: the built `wasm32-wasip1` executable plus its Arora header (JSON). Any Arora runtime can load it; in the browser, pass it to `@vizij/arora-web-wasm`'s `startDevice` so the device's engine hosts it and its functions (`load_animation`, `create_player`, `add_instance`, `step`, …) become callable.
+The vizij animation engine ([`vizij-animation-core`](https://github.com/vizij-ai/vizij-rs/tree/main/crates/animation/vizij-animation-core)) packaged as an **Arora wasm module**, shipped as importable assets: the built `wasm32-wasip1` executable plus its Arora header (JSON). Any Arora runtime can load it; in the browser, pass it to `@vizij/runtime`'s `startDevice` so the device's engine hosts it and its functions (`load_animation`, `create_player`, `add_instance`, `step`, …) become callable.
 
 ```ts
-import { startDevice } from "@vizij/arora-web-wasm";
+import { startDevice } from "@vizij/runtime";
 import { loadAnimationModule } from "@vizij/animation-module";
 
 const device = await startDevice(graph, undefined, [await loadAnimationModule()]);
