+++ package.json
@@ -29,13 +29,13 @@
     "link:wasm:orchestrator": "pnpm --filter @vizij/orchestrator-wasm run build && (cd npm/@vizij/orchestrator-wasm && pnpm link --global)",
     "clean:wasm:orchestrator": "rm -rf npm/@vizij/orchestrator-wasm/pkg npm/@vizij/orchestrator-wasm/dist",
     "build:wasm:arora-web": "node scripts/build-arora-web-wasm.mjs",
-    "link:wasm:arora-web": "pnpm --filter @vizij/arora-web-wasm run build && (cd npm/@vizij/arora-web-wasm && pnpm link --global)",
-    "clean:wasm:arora-web": "rm -rf npm/@vizij/arora-web-wasm/pkg npm/@vizij/arora-web-wasm/dist",
+    "link:wasm:arora-web": "pnpm --filter @vizij/runtime run build && (cd npm/@vizij/runtime && pnpm link --global)",
+    "clean:wasm:arora-web": "rm -rf npm/@vizij/runtime/pkg npm/@vizij/runtime/dist",
     "build:module:animation": "pnpm --filter @vizij/animation-module run build",
     "clean:module:animation": "pnpm --filter @vizij/animation-module run clean",
     "build:wasm": "node scripts/build-wasm.mjs",
     "clean:wasm": "pnpm run clean:wasm:animation && pnpm run clean:wasm:graph && pnpm run clean:wasm:orchestrator && pnpm run clean:wasm:arora-web",
-    "test:wasm": "pnpm --filter @vizij/animation-wasm run test && pnpm --filter @vizij/node-graph-wasm run test && pnpm --filter @vizij/orchestrator-wasm run test && pnpm --filter @vizij/animation-module run test && pnpm --filter @vizij/arora-web-wasm run test",
+    "test:wasm": "pnpm --filter @vizij/animation-wasm run test && pnpm --filter @vizij/node-graph-wasm run test && pnpm --filter @vizij/orchestrator-wasm run test && pnpm --filter @vizij/animation-module run test && pnpm --filter @vizij/runtime run test",
     "link:value-json": "pnpm --filter @vizij/value-json run build && (cd npm/@vizij/value-json && pnpm link --global)",
     "link:wasm": "pnpm run link:value-json && pnpm run link:wasm:animation && pnpm run link:wasm:graph && pnpm run link:wasm:orchestrator",
     "build:shared": "pnpm --filter @vizij/value-json --filter @vizij/wasm-loader --filter @vizij/test-fixtures run build",
