+++ scripts/build-arora-web-wasm.mjs
@@ -3,7 +3,7 @@ import { resolve } from "node:path";
 import { writeFileSync } from "node:fs";
 
 const crate = resolve(process.cwd(), "crates/interop/vizij-arora-web");
-const outDir = resolve(process.cwd(), "npm/@vizij/arora-web-wasm/pkg");
+const outDir = resolve(process.cwd(), "npm/@vizij/runtime/pkg");
 
 execSync(`wasm-pack build "${crate}" --target web --out-dir "${outDir}" --release`, {
   stdio: "inherit",
